+++ src/Feature/Board/BoardAction.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace voku\AgentUi\Feature\Board;
 
+use voku\AgentUi\Http\Request;
 use voku\AgentUi\Http\Response;
 use voku\AgentUi\Integration\AgentKanban\BoardProjectionGateway;
 use voku\AgentUi\View\TemplateRenderer;
@@ -16,8 +17,10 @@ final readonly class BoardAction
     ) {
     }
 
-    public function __invoke(): Response
+    public function __invoke(?Request $request = null): Response
     {
-        return Response::html($this->templates->render('board/index', ['board' => $this->board->board()]));
+        $boardId = $request?->query['board'] ?? null;
+
+        return Response::html($this->templates->render('board/index', ['board' => $this->board->board($boardId)]));
     }
 }